A Remake Of Pokemon 3D
Current Includes
1.Fuzzy Logic System (AI) Maxed Out For Most Trainers And Leaders

1.Wild Pokemon Changes
New Pokemon in The Wild Compared to the Last

New To Come In The Future In Complete Version
All Legends In Game
All Pokemon Stats Modified
All Moves Redone
All Trainers Pokemon Redone With Higher Level Difficulty
Wild Pokemon Will Have A Fuzzy Logic system (Will Need Help
An Added Language in English For Programming In Pokemon 3D. Will Need Help Making This Run (Need A Person Who Knows Programming To Help With This. Language is almost Done So i Need A compiler For It)
Plan To Add Animations If Language Gets Done As Well as Completely New System For the Game
Will Upgrade Textures For The Game too. Need a team For It, But If Done Will Look Amazing
And Finally Implement The Entire Game Into Unreal Engine and or Unity 3D Once Complete
And Much Much More To come Such As All Battle formats Ect